+++ CHANGELOG.md
@@ -22,6 +22,8 @@
   
 #### Enterprise Policies
 
+- new policy to allow or prevent Firefox from messaging the user, see
+  [#145](https://github.com/cadeyrn/enterprise-policy-generator/issues/145)
 - enhanced permissions policy to control autoplay of media and access to virtual reality devices, see
   [#136](https://github.com/cadeyrn/enterprise-policy-generator/issues/136)
 - enhanced tracking protection policy to add exceptions, see